#1552, Add source changes for posting per employee, reference to issue: https://github.com/adempiere/adempiere/issues/1552
