+++ base/src/org/eevolution/model/I_HR_Payroll.java
@@ -168,6 +168,19 @@ public interface I_HR_Payroll
 	  */
 	public boolean isIgnoreDefaultPayroll();
 
+    /** Column name IsPostPerEmployee */
+    public static final String COLUMNNAME_IsPostPerEmployee = "IsPostPerEmployee";
+
+	/** Set IsPostPerEmployee.
+	  * A Payroll is posted per employee when the value is Yes, it is posted accumulated when the value is N
+	  */
+	public void setIsPostPerEmployee (boolean IsPostPerEmployee);
+
+	/** Get IsPostPerEmployee.
+	  * A Payroll is posted per employee when the value is Yes, it is posted accumulated when the value is N
+	  */
+	public boolean isPostPerEmployee();
+
     /** Column name Name */
     public static final String COLUMNNAME_Name = "Name";
 
@@ -216,19 +229,6 @@ public interface I_HR_Payroll
 	/** Get Process Now	  */
 	public boolean isProcessing();
 
-    /** Column name UUID */
-    public static final String COLUMNNAME_UUID = "UUID";
-
-	/** Set Immutable Universally Unique Identifier.
-	  * Immutable Universally Unique Identifier
-	  */
-	public void setUUID (String UUID);
-
-	/** Get Immutable Universally Unique Identifier.
-	  * Immutable Universally Unique Identifier
-	  */
-	public String getUUID();
-
     /** Column name Updated */
     public static final String COLUMNNAME_Updated = "Updated";
 
@@ -244,6 +244,19 @@ public interface I_HR_Payroll
 	  * User who updated this records
 	  */
 	public int getUpdatedBy();
+
+    /** Column name UUID */
+    public static final String COLUMNNAME_UUID = "UUID";
+
+	/** Set Immutable Universally Unique Identifier.
+	  * Immutable Universally Unique Identifier
+	  */
+	public void setUUID (String UUID);
+
+	/** Get Immutable Universally Unique Identifier.
+	  * Immutable Universally Unique Identifier
+	  */
+	public String getUUID();
 
     /** Column name Value */
     public static final String COLUMNNAME_Value = "Value";
